requirements_new.txt
loguru==0.7.3
rich==13.9.4
openai==1.68.2
tenacity==8.2.3
httpx[socks]==0.27.2
pre-commit==3.6.0
pydantic-core==2.27.2
pydantic-xml==2.14.3
pydantic==2.10.6
z3-solver==4.14.1.0
pytest
pandas
pyyaml==6.0.2
ruamel.yaml==0.18.10
dill==0.3.9
anthropic==0.51.0
transformers==4.37.2
